backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
supabase==2.8.0
stripe==7.8.0
openai==1.3.5
requests==2.31.0
aiofiles==23.2.1
python-slugify==8.0.1
pydantic==2.5.0
email-validator==2.1.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
asyncpg==0.29.0
reportlab==4.0.7
Pillow==10.1.0
jinja2==3.1.2
httpx==0.24.1
websockets>=13.0
redis==5.0.1
celery==5.3.4
pydub==0.25.1
speechrecognition==3.10.0
TTS==0.20.6
ffmpeg-python==0.2.0